rename frag submodules to be hidden. This simplifies the public api
