+++ src/finaletoolkit/__init__.py
@@ -3,14 +3,11 @@ A package and standalone program to extract fragmentation features of
 cell-free DNA from paired-end sequencing data.
 """
 
-from finaletoolkit.version import __version__
+import lazy_loader as lazy
+
+from .version import __version__
 
 
-__all__ = ["cli", "frag", "genome", "utils"]
+subpackages = ["cli", "frag", "genome", "utils"]
 
-# Delay imports until the submodule is actually accessed.
-def __getattr__(name):
-    if name in __all__:
-        import importlib
-        return importlib.import_module(f".{name}", __name__)
-    raise AttributeError(f"Module {__name__} has no attribute {name}")
+__getattr__, __dir__, _ = lazy.attach(__name__, subpackages)
