+++ src/finaletoolkit/frag/__init__.py
@@ -1,10 +1,10 @@
-from finaletoolkit.frag.frag_length import frag_length, frag_length_bins
-from finaletoolkit.frag.coverage import coverage
-from finaletoolkit.frag.wps import wps
-from finaletoolkit.frag.multi_wps import multi_wps
-from finaletoolkit.frag.cleavage_profile import cleavage_profile
-from finaletoolkit.frag.delfi import delfi
-from finaletoolkit.frag.delfi_gc_correct import delfi_gc_correct
-from finaletoolkit.frag.delfi_merge_bins import delfi_merge_bins
-from finaletoolkit.frag.adjust_wps import adjust_wps
-from finaletoolkit.frag.end_motifs import *
+from finaletoolkit.frag._frag_length import frag_length, frag_length_bins, frag_length_intervals
+from finaletoolkit.frag._coverage import coverage, single_coverage
+from finaletoolkit.frag._wps import wps
+from finaletoolkit.frag._multi_wps import multi_wps
+from finaletoolkit.frag._cleavage_profile import cleavage_profile
+from finaletoolkit.frag._delfi import delfi
+from finaletoolkit.frag._delfi_gc_correct import delfi_gc_correct
+from finaletoolkit.frag._delfi_merge_bins import delfi_merge_bins
+from finaletoolkit.frag._adjust_wps import adjust_wps
+from finaletoolkit.frag._end_motifs import *
